+++ src/SSB_RetuningAutomations.py
@@ -43,7 +43,7 @@ from src.modules.CleanUp.FinalCleanUp import FinalCleanUp
 
 TOOL_NAME           = "SSB_RetuningAutomations"
 TOOL_VERSION        = "0.5.1"
-TOOL_DATE           = "2025-12-22"
+TOOL_DATE           = "2025-12-23"
 TOOL_NAME_VERSION   = f"{TOOL_NAME}_v{TOOL_VERSION}"
 COPYRIGHT_TEXT      = "(c) 2025 - Jaime Tur (jaime.tur@ericsson.com)"
 TOOL_DESCRIPTION    = textwrap.dedent(f"""